move files to algebra dir
