+++ algebra/src/config.rs
