+++ algebra/src/msm.rs
