--- crypto/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod cyclic_group;
-pub mod elliptic_curve;
-pub mod field_element;
-pub mod field_extension_element;
-pub mod msm;
-pub mod polynomial;
-pub mod config;
